Projet fini

+++ backend/controllers/books.js
@@ -55,32 +55,50 @@ exports.getOneBook = (req, res, next) => {
 
 // --- Modifier un livre avec son id ---
 exports.modifyOneBook = (req, res, next) => {
-  const bookObject = req.file
-    ? {
-        ...JSON.parse(req.body.book),
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${
-          req.file.filename
-        }`,
-      }
-    : { ...req.body };
-  delete bookObject._userId;
-  Book.findOne({ _id: req.params.id })
-    .then((book) => {
-      if (book.userId != req.auth.userId) {
-        res.status(401).json({ message: 'Non-autorisé !' });
-      } else {
-        const filename = book.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
+  let bookObject;
+  if (req.file) {
+    bookObject = {
+      ...JSON.parse(req.body.book),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${
+        req.file.filename
+      }`,
+    };
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+      .then((book) => {
+        if (book.userId != req.auth.userId) {
+          res.status(401).json({ message: 'Non-autorisé !' });
+        } else {
+          const filename = book.imageUrl.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {
+            Book.updateOne(
+              { _id: req.params.id },
+              { ...bookObject, _id: req.params.id }
+            )
+              .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+              .catch((error) => res.status(400).json({ error }));
+          });
+        }
+      })
+      .catch((error) => res.status(400).json({ error }));
+  } else {
+    bookObject = { ...req.body };
+    delete bookObject._userId;
+    Book.findOne({ _id: req.params.id })
+      .then((book) => {
+        if (book.userId != req.auth.userId) {
+          res.status(401).json({ message: 'Non-autorisé !' });
+        } else {
           Book.updateOne(
             { _id: req.params.id },
             { ...bookObject, _id: req.params.id }
           )
             .then(() => res.status(200).json({ message: 'Objet modifié !' }))
             .catch((error) => res.status(400).json({ error }));
-        });
-      }
-    })
-    .catch((error) => res.status(400).json({ error }));
+        }
+      })
+      .catch((error) => res.status(400).json({ error }));
+  }
 };
 
 // --- Supprimer un livre avec son id ---
